added createConfigFile functionality
